requirements.txt
crawl4ai>=0.8.0
httpx>=0.27.0
python-dateutil>=2.8.0
rich>=13.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
respx>=0.21.0
claude-agent-sdk>=0.1.0
